# docs/api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DocReader PoC API",
    "description": "Local proof of concept for document ingestion, OCR, classification and extraction.\n\n**There is no authentication.** Every endpoint is anonymous, the Swagger page is public and\nno bearer token is required. Do not expose this API to the internet and do not send real\npersonal documents to it. Use synthetic, masked or authorized files only.\n\nConventions: JSON in camelCase, timestamps in UTC ISO 8601, errors as\n`application/problem+json`, and an `X-Correlation-Id` header on every response.\n\nStage 3 of the execution plan delivers upload, listing, detail, content, deletion, OCR\n(PP-OCRv5 on CPU, see ADR 0002), the raw text and the structured result of seven document\ntypes: CPF card, CIN/RG, CNH, proof of address, CNPJ card, CCMEI and social contract.\nOther document types are read as raw text only.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/product-services": {
      "post": {
        "tags": [
          "ProductServices"
        ],
        "summary": "Creates a product or service.",
        "description": "The code is unique without regard to case, is stored in upper case and cannot be changed later.",
        "requestBody": {
          "description": "Code, name and active flag.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductServiceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductServiceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductServiceRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/product-services",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "201": {
            "description": "Created. The `Location` header points at it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "The code or the name is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/product-services",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "409": {
            "description": "The code is already taken.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "A product or service with the code CONTA-PJ already exists.",
                  "instance": "/api/v1/product-services",
                  "errorCode": "PRODUCT_SERVICE_CODE_EXISTS",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The storage repository does not exist or is inactive.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "The document has 84 pages and the limit is 50 pages.",
                  "instance": "/api/v1/product-services",
                  "errorCode": "PAGE_LIMIT_EXCEEDED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ProductServices"
        ],
        "summary": "Lists the products and services, by code.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/product-services",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "One page of products and services.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductServiceResponsePagedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/product-services/{id}": {
      "get": {
        "tags": [
          "ProductServices"
        ],
        "summary": "Returns one product or service.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "The product or service.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductServiceResponse"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No product service matches the requested identifier.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "PRODUCT_SERVICE_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductServices"
        ],
        "summary": "Changes the name, the active flag and the storage repository of a product or service.",
        "description": "The code is immutable. `storageRepositoryId` replaces the current one (omit it to go back to the default repository); documents already stored stay where they are. Deactivating a product does not touch its documents; it refuses new uploads.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "New name and active flag.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductServiceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductServiceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductServiceRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "The name is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No product service matches the requested identifier.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "PRODUCT_SERVICE_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The storage repository does not exist or is inactive.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "The document has 84 pages and the limit is 50 pages.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "PAGE_LIMIT_EXCEEDED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProductServices"
        ],
        "summary": "DESTRUCTIVE AND IRREVERSIBLE — Deletes a product or service that nothing refers to.",
        "description": "A product with documents, retention policies or webhook subscriptions cannot be deleted: deactivate it instead.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "204": {
            "description": "Deleted."
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No product service matches the requested identifier.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "PRODUCT_SERVICE_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "409": {
            "description": "Something still refers to it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "Documents, retention policies or webhook subscriptions still refer to this product or service. Deactivate it (active = false) instead of deleting it.",
                  "instance": "/api/v1/product-services/{id}",
                  "errorCode": "PRODUCT_SERVICE_IN_USE",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProductServiceRequest": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string",
            "description": "Unique code: 1 to 64 letters, digits, dots, hyphens or underscores, starting with a letter or digit. Stored in upper case and immutable."
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "storageRepositoryId": {
            "type": "string",
            "description": "Repository the documents of this product are stored in; omit to use the default repository. Must exist and be active.",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProductServiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique code, in upper case. Send it as `productServiceCode` when uploading."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "active": {
            "type": "boolean",
            "description": "An inactive product keeps its documents but refuses new uploads."
          },
          "storageRepository": {
            "$ref": "#/components/schemas/StorageRepositoryReferenceResponse"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation instant, in UTC.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last change, in UTC.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "A product or service documents can be linked to.",
        "example": {
          "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
          "code": "CONTA-PJ",
          "name": "Abertura de conta PJ",
          "active": true,
          "storageRepository": {
            "id": "0199c1f0-3333-7a10-9c44-2f1d8e6b4a21",
            "code": "ARQUIVO-DB",
            "name": "Arquivo no banco de dados"
          },
          "createdAt": "2026-09-24T22:00:00Z",
          "updatedAt": "2026-09-24T22:00:00Z"
        }
      },
      "ProductServiceResponsePagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductServiceResponse"
            },
            "description": "Items of the current page, newest first."
          },
          "page": {
            "type": "integer",
            "description": "One based page number.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of items matching the filters.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "Number of pages for the current page size.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Envelope of every paginated listing.",
        "example": {
          "items": [
            {
              "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
              "code": "CONTA-PJ",
              "name": "Abertura de conta PJ",
              "active": true,
              "storageRepository": {
                "id": "0199c1f0-3333-7a10-9c44-2f1d8e6b4a21",
                "code": "ARQUIVO-DB",
                "name": "Arquivo no banco de dados"
              },
              "createdAt": "2026-09-24T22:00:00Z",
              "updatedAt": "2026-09-24T22:00:00Z"
            }
          ],
          "page": 1,
          "pageSize": 20,
          "totalCount": 1,
          "totalPages": 1
        }
      },
...
      "StorageRepositoryReferenceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique code."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        },
        "additionalProperties": false,
        "description": "The storage repository a product or service stores its documents in."
      },
...
      "UpdateProductServiceRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "storageRepositoryId": {
            "type": "string",
            "description": "Repository the documents of this product are stored in; omit (null) to use the default repository. Must exist and be active.",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}